+++ src/scripts/app/tidy5e-lazy-exp-and-hp.js
@@ -68,6 +68,8 @@ function _onChangeExp(ev) {
 		newAmountExp = 0;
 	}
 
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeExp | Update system.details.xp.value to ${newAmountExp}`);
+
 	sheet.submitOnChange = false;
 	actor
 		.update({ "system.details.xp.value": Number(newAmountExp) })
@@ -157,6 +159,10 @@ function _onChangeHp(ev) {
 			newAmountHpTempMax = 0;
 		}
 
+    debug(`tidy5e-lazy-exp-and-hp | _onChangeHp | Update system.attributes.hp.value to ${newAmountHpValue}`);
+    debug(`tidy5e-lazy-exp-and-hp | _onChangeHp | Update system.attributes.hp.temp to ${newAmountHpTemp}`);
+    debug(`tidy5e-lazy-exp-and-hp | _onChangeHp | Update system.attributes.hp.tempmax to ${newAmountHpTempMax}`);
+
 		sheet.submitOnChange = false;
 		actor
 			.update({
@@ -180,6 +186,8 @@ function _onChangeHp(ev) {
 				newAmountHpValue = maxHp + newAmountHpTempMax;
 			}
 		}
+
+    debug(`tidy5e-lazy-exp-and-hp | _onChangeHp | Update system.attributes.hp.value to ${newAmountHpValue}`);
 
 		sheet.submitOnChange = false;
 		actor
@@ -251,6 +259,8 @@ function _onChangeHpMax(ev) {
 		newAmountHpMax = 0;
 	}
 
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeHpMax | Update system.attributes.hp.max to ${newAmountHpMax}`);
+
 	sheet.submitOnChange = false;
 	actor
 		.update({ "system.attributes.hp.max": Number(newAmountHpMax) })
@@ -293,6 +303,8 @@ function _onChangeHpForceHpValueLimit1(ev) {
 	if (newAmountHpValue > maxHp + newAmountHpTempMax) {
 		newAmountHpValue = maxHp + newAmountHpTempMax;
 	}
+
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeHpForceHpValueLimit1 | Update system.attributes.hp.value to ${newAmountHpValue}`);
 
 	sheet.submitOnChange = false;
 	actor
@@ -349,6 +361,10 @@ function _onChangeHpForceHpValueLimit2(ev) {
 		newAmountHpTempMax = 0;
 	}
 
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeHpForceHpValueLimit2 | Update system.attributes.hp.value to ${newAmountHpValue}`);
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeHpForceHpValueLimit2 | Update system.attributes.hp.temp to ${newAmountHpTemp}`);
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeHpForceHpValueLimit2 | Update system.attributes.hp.tempmax to ${newAmountHpTempMax}`);
+
 	sheet.submitOnChange = false;
 	actor
 		.update({
@@ -381,6 +397,8 @@ function _onChangeHpForceHpTempLimit2(ev) {
 		debug(`tidy5e-lazy-exp-and-hp | _onChangeHpForceHpTempLimit2 | [18] WARN: The hp.temp value ${newAmountHpTemp} is not a valid number`);
 		newAmountHpTemp = 0;
 	}
+
+  debug(`tidy5e-lazy-exp-and-hp | _onChangeHpForceHpTempLimit2 | Update system.attributes.hp.temp to ${newAmountHpTemp}`);
 
 	sheet.submitOnChange = false;
 	actor
@@ -497,99 +515,99 @@ export function applyLazyHp(app, html, actorData) {
 	}
 }
 
-Hooks.on("preUpdateActor", function (actorEntity, update, options, userId) {
-	if (!game.settings.get(CONSTANTS.MODULE_ID, "lazyHpAndExpEnable")) {
-		return;
-	}
+// Hooks.on("preUpdateActor", function (actorEntity, update, options, userId) {
+// 	if (!game.settings.get(CONSTANTS.MODULE_ID, "lazyHpAndExpEnable")) {
+// 		return;
+// 	}
 
-	if (!actorEntity) {
-		return;
-	}
-	if (hasProperty(update, "system.attributes.hp.value")) {
-		let hpValue = getProperty(update, "system.attributes.hp.value");
-		if(isEmptyObject(hpValue)) {
-			// Do nothing
-		} else {
-			if (!is_lazy_number(hpValue) && Number(hpValue) !== Number(actorEntity.system.attributes.hp.value)) {
-				setProperty(update, "system.attributes.hp.value", Number(hpValue));
-				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [0] update system.attributes.hp.value from '${hpValue}' to '${getProperty(update, "system.attributes.hp.value")}'`);
-			}
-			// Module compatibility with https://foundryvtt.com/packages/link-item-resource-5e
-			else if (String(hpValue).startsWith("0") && String(hpValue) !== "0") {
-				let hpValueTmp = hpValue;
-				while (String(hpValueTmp).startsWith("0")) {
-					if (String(hpValueTmp) === "0") {
-						break;
-					}
-					hpValueTmp = String(hpValueTmp).slice(1);
-				}
-				if(Number(hpValueTmp) !== Number(actorEntity.system.attributes.hp.value)) {
-					setProperty(update, "system.attributes.hp.value", Number(hpValueTmp ?? 0));
-					info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [1] update system.attributes.hp.value from '${hpValue}' to '${getProperty(update, "system.attributes.hp.value")}'`);
-				}
-			}
-			if(!is_real_number(getProperty(update, "system.attributes.hp.value"))) {
-				setProperty(update, "system.attributes.hp.value", 0);
-				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [2] update system.attributes.hp.value from '${getProperty(update, "system.attributes.hp.value")}' to number 0`);
-			}
-		}
-	}
-	if (hasProperty(update, "system.attributes.hp.max")) {
-		let hpMaxValue = getProperty(update, "system.attributes.hp.max");
-		if(isEmptyObject(hpMaxValue)) {
-			// Do nothing
-		} else {
-			if (!is_lazy_number(hpMaxValue) && Number(hpMaxValue) !== Number(actorEntity.system.attributes.hp.max)) {
-				setProperty(update, "system.attributes.hp.max", Number(hpMaxValue ?? 0));
-				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [0] update system.attributes.hp.max from '${hpMaxValue}' to '${getProperty(update, "system.attributes.hp.max")}'`);
-			}
-			// Module compatibility with https://foundryvtt.com/packages/link-item-resource-5e
-			else if (String(hpMaxValue).startsWith("0") && String(hpMaxValue) !== "0") {
-				let hpMaxValueTmp = hpMaxValue;
-				while (String(hpMaxValueTmp).startsWith("0")) {
-					if (String(hpMaxValueTmp) === "0") {
-						break;
-					}
-					hpMaxValueTmp  = String(hpMaxValueTmp).slice(1);
-				}
-				if(Number(hpMaxValueTmp) !== Number(actorEntity.system.attributes.hp.max)) {
-					setProperty(update, "system.attributes.hp.max", Number(hpMaxValueTmp ?? 0));
-					info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [1] update system.attributes.hp.max from '${hpMaxValue}' to '${getProperty(update, "system.attributes.hp.max")}'`);
-				}
-			}
-			if(!is_real_number(getProperty(update, "system.attributes.hp.max"))) {
-				setProperty(update, "system.attributes.hp.max", 0);
-				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [2] update system.attributes.hp.max from '${getProperty(update, "system.attributes.hp.max")}' to number 0`);
-			}
-		}
-	}
-	if (hasProperty(update, "system.details.xp.value")) {
-		let xpValue = getProperty(update, "system.details.xp.value");
-		if(isEmptyObject(xpValue)) {
-			// Do nothing
-		} else {
-			if (!is_lazy_number(xpValue) && Number(xpValue) !== Number(actorEntity.system.details.xp.value)) {
-				setProperty(update, "system.details.xp.value", Number(xpValue ?? 0));
-				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [0] update system.details.xp.value from '${xpValue}' to '${getProperty(update, "system.details.xp.value")}'`);
-			}
-			// Module compatibility with https://foundryvtt.com/packages/link-item-resource-5e
-			else if (String(xpValue).startsWith("0") && String(xpValue) !== "0") {
-				let xpValueTmp = xpValue;
-				while (String(xpValueTmp).startsWith("0")) {
-					if (String(xpValueTmp) === "0") {
-						break;
-					}
-					xpValueTmp = String(xpValueTmp).slice(1);
-				}
-				if(Number(xpValueTmp) !== Number(actorEntity.system.details.xp.value)) {
-					setProperty(update, "system.details.xp.value", Number(xpValueTmp ?? 0));
-					info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [1] update system.details.xp.value from '${xpValue}' to '${getProperty(update, "system.details.xp.value")}'`);
-				}
-			}
-			if(!is_real_number(getProperty(update, "system.details.xp.value"))) {
-				setProperty(update, "system.details.xp.value", 0);
-				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [2] update system.details.xp.value from '${getProperty(update, "system.details.xp.value")}' to number 0`);
-			}
-		}
-	}
-});
+// 	if (!actorEntity) {
+// 		return;
+// 	}
+// 	if (hasProperty(update, "system.attributes.hp.value")) {
+// 		let hpValue = getProperty(update, "system.attributes.hp.value");
+// 		if(isEmptyObject(hpValue)) {
+// 			// Do nothing
+// 		} else {
+// 			if (!is_lazy_number(hpValue) && Number(hpValue) !== Number(actorEntity.system.attributes.hp.value)) {
+// 				setProperty(update, "system.attributes.hp.value", Number(hpValue));
+// 				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [0] update system.attributes.hp.value from '${hpValue}' to '${getProperty(update, "system.attributes.hp.value")}'`);
+// 			}
+// 			// Module compatibility with https://foundryvtt.com/packages/link-item-resource-5e
+// 			else if (String(hpValue).startsWith("0") && String(hpValue) !== "0") {
+// 				let hpValueTmp = hpValue;
+// 				while (String(hpValueTmp).startsWith("0")) {
+// 					if (String(hpValueTmp) === "0") {
+// 						break;
+// 					}
+// 					hpValueTmp = String(hpValueTmp).slice(1);
+// 				}
+// 				if(Number(hpValueTmp) !== Number(actorEntity.system.attributes.hp.value)) {
+// 					setProperty(update, "system.attributes.hp.value", Number(hpValueTmp ?? 0));
+// 					info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [1] update system.attributes.hp.value from '${hpValue}' to '${getProperty(update, "system.attributes.hp.value")}'`);
+// 				}
+// 			}
+// 			if(!is_real_number(getProperty(update, "system.attributes.hp.value"))) {
+// 				setProperty(update, "system.attributes.hp.value", 0);
+// 				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [2] update system.attributes.hp.value from '${getProperty(update, "system.attributes.hp.value")}' to number 0`);
+// 			}
+// 		}
+// 	}
+// 	if (hasProperty(update, "system.attributes.hp.max")) {
+// 		let hpMaxValue = getProperty(update, "system.attributes.hp.max");
+// 		if(isEmptyObject(hpMaxValue)) {
+// 			// Do nothing
+// 		} else {
+// 			if (!is_lazy_number(hpMaxValue) && Number(hpMaxValue) !== Number(actorEntity.system.attributes.hp.max)) {
+// 				setProperty(update, "system.attributes.hp.max", Number(hpMaxValue ?? 0));
+// 				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [0] update system.attributes.hp.max from '${hpMaxValue}' to '${getProperty(update, "system.attributes.hp.max")}'`);
+// 			}
+// 			// Module compatibility with https://foundryvtt.com/packages/link-item-resource-5e
+// 			else if (String(hpMaxValue).startsWith("0") && String(hpMaxValue) !== "0") {
+// 				let hpMaxValueTmp = hpMaxValue;
+// 				while (String(hpMaxValueTmp).startsWith("0")) {
+// 					if (String(hpMaxValueTmp) === "0") {
+// 						break;
+// 					}
+// 					hpMaxValueTmp  = String(hpMaxValueTmp).slice(1);
+// 				}
+// 				if(Number(hpMaxValueTmp) !== Number(actorEntity.system.attributes.hp.max)) {
+// 					setProperty(update, "system.attributes.hp.max", Number(hpMaxValueTmp ?? 0));
+// 					info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [1] update system.attributes.hp.max from '${hpMaxValue}' to '${getProperty(update, "system.attributes.hp.max")}'`);
+// 				}
+// 			}
+// 			if(!is_real_number(getProperty(update, "system.attributes.hp.max"))) {
+// 				setProperty(update, "system.attributes.hp.max", 0);
+// 				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [2] update system.attributes.hp.max from '${getProperty(update, "system.attributes.hp.max")}' to number 0`);
+// 			}
+// 		}
+// 	}
+// 	if (hasProperty(update, "system.details.xp.value")) {
+// 		let xpValue = getProperty(update, "system.details.xp.value");
+// 		if(isEmptyObject(xpValue)) {
+// 			// Do nothing
+// 		} else {
+// 			if (!is_lazy_number(xpValue) && Number(xpValue) !== Number(actorEntity.system.details.xp.value)) {
+// 				setProperty(update, "system.details.xp.value", Number(xpValue ?? 0));
+// 				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [0] update system.details.xp.value from '${xpValue}' to '${getProperty(update, "system.details.xp.value")}'`);
+// 			}
+// 			// Module compatibility with https://foundryvtt.com/packages/link-item-resource-5e
+// 			else if (String(xpValue).startsWith("0") && String(xpValue) !== "0") {
+// 				let xpValueTmp = xpValue;
+// 				while (String(xpValueTmp).startsWith("0")) {
+// 					if (String(xpValueTmp) === "0") {
+// 						break;
+// 					}
+// 					xpValueTmp = String(xpValueTmp).slice(1);
+// 				}
+// 				if(Number(xpValueTmp) !== Number(actorEntity.system.details.xp.value)) {
+// 					setProperty(update, "system.details.xp.value", Number(xpValueTmp ?? 0));
+// 					info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [1] update system.details.xp.value from '${xpValue}' to '${getProperty(update, "system.details.xp.value")}'`);
+// 				}
+// 			}
+// 			if(!is_real_number(getProperty(update, "system.details.xp.value"))) {
+// 				setProperty(update, "system.details.xp.value", 0);
+// 				info(`tidy5e-lazy-exp-and-hp | preUpdateActor | [2] update system.details.xp.value from '${getProperty(update, "system.details.xp.value")}' to number 0`);
+// 			}
+// 		}
+// 	}
+// });