+++ changelog.md
@@ -1,3 +1,7 @@
+### 0.8.31
+
+- Remove preUpdateActor from 'lazyhp' feature for https://github.com/sdenec/tidy5e-sheet/issues/696 and https://github.com/sdenec/tidy5e-sheet/issues/707
+- Separate 'allowHpMaxOverride' in 'allowHpMaxOverride' and 'allowHpConfigOverride'
 ### 0.8.28-29-30
 
 - Fix import js